Move the runnable routine to the BackUpRepository
