+++ app/src/main/java/com/github/factotum_sdp/factotum/repositories/BackUpRepository.kt
@@ -1,7 +1,10 @@
 package com.github.factotum_sdp.factotum.repositories
 
+import android.os.Handler
+import android.os.Looper
 import androidx.datastore.core.DataStore
 import com.github.factotum_sdp.factotum.firebase.FirebaseInstance
+import com.github.factotum_sdp.factotum.ui.roadbook.RoadBookViewModel
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.DatabaseReference
@@ -23,6 +26,26 @@ abstract class BackUpRepository<T: List<Any>>(remoteSource: DatabaseReference, u
 
     private var lastNetworkBackUp: T? = null
     private var lastLocalBackUp: T? = null
+
+    private val handler = Handler(Looper.getMainLooper())
+    private val fetchDataRunnable = object : Runnable {
+        override fun run() {
+            //roadBookRepository.setBackUp(currentDRecList())
+            handler.postDelayed(this, WAIT_TIME_BACK_UP_UPDATE)
+        }
+    }
+
+    fun launchRunnableBackUp() {
+        if(withTimedBackUp) {
+            handler.post(fetchDataRunnable)
+        }
+    }
+
+    fun clearRunnableBackUp() {
+        if(withTimedBackUp) {
+            handler.removeCallbacks(fetchDataRunnable)
+        }
+    }
 
     init {
         FirebaseInstance.onConnectedStatusChanged {
@@ -106,5 +129,13 @@ abstract class BackUpRepository<T: List<Any>>(remoteSource: DatabaseReference, u
 
     private suspend fun getLastLocalBackUp(): T {
         return lastLocalBackUp ?: localSource.data.first()
+    }
+
+    companion object {
+        const val WAIT_TIME_BACK_UP_UPDATE = 15000L
+        var withTimedBackUp = true
+        fun setTimedBackUp(isEnabled: Boolean) { // For testing purpose
+            withTimedBackUp = isEnabled
+        }
     }
 }
